+++ reasoner_validator/__init__.py
@@ -1,9 +1,12 @@
 from typing import Optional, List, Dict
+
 from reasoner_validator.report import ValidationReporter
 from reasoner_validator.biolink import (
     check_biolink_model_compliance_of_query_graph,
     check_biolink_model_compliance_of_knowledge_graph,
-    BiolinkValidator
+    BiolinkValidator,
+    get_biolink_model_toolkit,
+    BiolinkValidator, TRAPIGraphType
 )
 
 # Maximum number of data points to scrutinize
@@ -161,7 +164,7 @@ class TRAPIResponseValidator(ValidationReporter):
         :param graph: original knowledge graph
         :type graph: Dict
         :param edges_limit: integer maximum number of edges to be validated in the knowledge graph. A value of zero
-                            triggers validation of all edges in the knowledge graph (could take awhile! Default: 100)
+                            triggers validation of all edges in the knowledge graph (could take a while! Default: 100)
         :type edges_limit: int
 
         :return: Dict, 'edges_limit' sized subset of knowledge graph
@@ -177,7 +180,7 @@ class TRAPIResponseValidator(ValidationReporter):
             # The kg_sample size will always be the 'sample_size'
             # or less, the latter situation arising if the number
             # of graph edges is smaller or some subject or
-            # object id's are missing in the nodes list.
+            # object ids are missing in the nodes list.
             sample_size = min(edges_limit, len(graph["edges"]))
             n = 0
             for key, edge in graph['edges'].items():
@@ -258,7 +261,7 @@ class TRAPIResponseValidator(ValidationReporter):
         :param message: input message expected to contain the 'knowledge_graph'
         :type message: Dict
         :param edges_limit: integer maximum number of edges to be validated in the knowledge graph. A value of zero
-                            triggers validation of all edges in the knowledge graph (could take awhile! Default: 100)
+                            triggers validation of all edges in the knowledge graph (could take a while! Default: 100)
         :type edges_limit: int
 
         :return: bool, False, if validation errors
@@ -412,3 +415,336 @@ class TRAPIResponseValidator(ValidationReporter):
 
         # Only 'error' but not 'info' nor 'warning' messages invalidate the overall Message
         return False if self.has_errors() else True
+
+    @staticmethod
+    def case_node_found(target: str, identifier: str, case: Dict, nodes: Dict) -> bool:
+        """
+        Check for presence of the target identifier,
+        with expected categories, in the "nodes" catalog.
+
+        :param target: 'subject' or 'object'
+        :param identifier: (CURIE) identifier of the node
+        :param case: Dict, full test case (to access the target node 'category')
+        :param nodes: Dict, nodes category indexed by node identifiers.
+        :return:
+        """
+        #
+        #     "nodes": {
+        #         "MONDO:0005148": {"name": "type-2 diabetes"},
+        #         "CHEBI:6801": {"name": "metformin", "categories": ["biolink:Drug"]}
+        #     }
+        #
+
+        # Sanity check
+        assert target in ["subject", "object"]
+
+        if identifier in nodes.keys():
+            # Found the target node identifier,
+            # but is the expected category present?
+            node_details = nodes[identifier]
+            if "categories" in node_details:
+                category = case[f"{target}_category"]
+                if category in node_details["categories"]:
+                    return True
+
+        # Target node identifier or categories is missing,
+        # or not annotated with the expected category?
+        return False
+
+    @staticmethod
+    def case_edge_bindings(target_edge_id: str, data: Dict) -> bool:
+        """
+        Check if target query edge id and knowledge graph edge id are in specified edge_bindings.
+        :param target_edge_id:  str, expected knowledge edge identifier in a matching result
+        :param data: TRAPI version-specific Response context from which the 'edge_bindings' may be retrieved
+        :return: True, if found
+        """
+        edge_bindings: Dict = data["edge_bindings"]
+        for bound_query_id, edge in edge_bindings.items():
+            # The expected query identifier in this context is
+            # hard coded as 'ab' in the 'one_hop.util.py' model
+            if bound_query_id == "ab":
+                for binding_details in edge:
+                    # TRAPI schema validation actually
+                    # catches missing id's, but sanity check...
+                    if "id" in binding_details:
+                        if target_edge_id == binding_details["id"]:
+                            return True
+        return False
+
+    def case_result_found(
+            self,
+            subject_id: str,
+            object_id: str,
+            edge_id: str,
+            results: List,
+            trapi_version: str
+    ) -> bool:
+        """
+        Validate that test case S--P->O edge is found bound to the Results?
+        :param subject_id: str, subject node (CURIE) identifier
+        :param object_id:  str, subject node (CURIE) identifier
+        :param edge_id:  str, subject node (CURIE) identifier
+        :param results: List of (TRAPI-version specific) Result objects
+        :param trapi_version: str, target TRAPI version of the Response being validated
+        :return: bool, True if case S-P-O edge was found in the results
+        """
+        trapi_1_4_0: bool
+        try:    # try block ... Sanity check: in case the trapi_version is somehow invalid?
+            target_version: SemVer = SemVer.from_string(trapi_version)
+            trapi_1_4_0 = target_version >= SemVer.from_string("1.4.0")
+        except SemVerError as sve:
+            logger.warning(f"case_result_found() 'trapi_version' seems invalid: {str(sve)}. Default to latest?")
+            trapi_1_4_0 = True
+
+        result_found: bool = False
+        result: Dict
+
+        for result in results:
+
+            # Node binding validation still currently same for recent TRAPI versions
+            node_bindings: Dict = result["node_bindings"]
+            subject_id_found: bool = False
+            object_id_found: bool = False
+            edge_id_found: bool = False
+            for node in node_bindings.values():
+                for details in node:
+                    if "id" in details:
+                        if subject_id == details["id"]:
+                            subject_id_found = True
+                        elif object_id == details["id"]:
+                            object_id_found = True
+
+            # However, TRAPI 1.4.0 Message 'Results' 'edge_bindings' are reported differently
+            #          from 1.3.0, rather, embedded in 'Analysis' objects (and 'Auxiliary Graphs')
+            if trapi_1_4_0:
+                #
+                #     "auxiliary_graphs": {
+                #         "a0": {
+                #             "edges": [
+                #                 "e02",
+                #                 "e12"
+                #             ]
+                #         },
+                #         "a1": {
+                #             "edges": [
+                #                 "extra_edge0"
+                #             ]
+                #         },
+                #         "a2": {
+                #             "edges" [
+                #                 "extra_edge1"
+                #             ]
+                #         }
+                #     },
+                #     "results": [
+                #         {
+                #             "node_bindings": {
+                #                 "n0": [
+                #                     "id": "diabetes"
+                #                 ],
+                #                 "n1": [
+                #                     "id": "metformin"
+                #                 ]
+                #             },
+                #             "analyses":[
+                #                 {
+                #                     "reasoner_id": "ara0",
+                #                     "edge_bindings": {
+                #                         "e0": [
+                #                             {
+                #                                 "id": "e01"
+                #                             },
+                #                             {
+                #                                 "id": "creative_edge"
+                #                             }
+                #                         ]
+                #                     },
+                #                     "support_graphs": [
+                #                         "a1",
+                #                         "a2"
+                #                     ]
+                #                     "score": ".7"
+                #                 },
+                #             ]
+                #         }
+                #     ]
+
+                # result["analyses"] may be empty but prior TRAPI 1.4.0 schema validation ensures that
+                # the "analysis" key is at least present plus the objects themselves are 'well-formed'
+                analyses: List = result["analyses"]
+                for analysis in analyses:
+                    edge_id_found = self.case_edge_bindings(edge_id, analysis)
+                    if edge_id_found:
+                        break
+
+            else:
+                # TRAPI 1.3.0 or earlier?
+                #
+                # Then, the TRAPI 1.3.0 Message Results (referencing the
+                # Response Knowledge Graph) could be something like this:
+                #
+                #     "results": [
+                #         {
+                #             "node_bindings": {
+                #                # node "id"'s in knowledge graph, in edge "id"
+                #                 "type-2 diabetes": [{"id": "MONDO:0005148"}],
+                #                 "drug": [{"id": "CHEBI:6801"}]
+                #             },
+                #             "edge_bindings": {
+                #                 # the edge binding key should be the query edge id
+                #                 # bounded edge "id" is from knowledge graph
+                #                 "treats": [{"id": "df87ff82"}]
+                #             }
+                #         }
+                #     ]
+                #
+                edge_id_found = self.case_edge_bindings(edge_id, result)
+
+            if subject_id_found and object_id_found and edge_id_found:
+                result_found = True
+                break
+
+        return result_found
+
+    def case_input_found_in_response(
+            self,
+            case: Dict,
+            response: Dict,
+            trapi_version: str
+    ) -> bool:
+        """
+        Predicate to validate if test data test case specified edge is returned
+        in the Knowledge Graph of the TRAPI Response Message. This method assumes
+        that the TRAPI response is already generally validated as well-formed.
+
+        :param case: Dict, input data test case
+        :param response: Dict, TRAPI Response whose message ought to contain the test case edge
+        :param trapi_version: str, TRAPI version of response being tested
+        :return: True if test case edge found; False otherwise
+        """
+        # sanity checks
+        assert case, "case_input_found_in_response(): Empty or missing test case data!"
+        assert response, "case_input_found_in_response(): Empty or missing TRAPI Response!"
+        assert "message" in response, "case_input_found_in_response(): TRAPI Response missing its Message component!"
+        assert trapi_version
+
+        #
+        # case: Dict parameter contains something like:
+        #
+        #     idx: 0,
+        #     subject_category: 'biolink:SmallMolecule',
+        #     object_category: 'biolink:Disease',
+        #     predicate: 'biolink:treats',
+        #     subject_id: 'CHEBI:3002',  # may have the deprecated key 'subject' here
+        #     object_id: 'MESH:D001249', # may have the deprecated key 'object' here
+        #
+        # the contents for which ought to be returned in
+        # the TRAPI Knowledge Graph, as a Result mapping?
+        #
+
+        message: Dict = response["message"]
+        if not (
+            "knowledge_graph" in message and message["knowledge_graph"] and
+            "results" in message and message["results"]
+        ):
+            # empty knowledge graph is syntactically ok, but in
+            # this, input test data edge is automatically deemed missing
+            return False
+
+        # TODO: We need to check **here*** whether or not the
+        #       TRAPI response returned the original test case edge!!?!!
+        #       Not totally sure if we should first search the Results then
+        #       the Knowledge Graph, or go directly to the Knowledge Graph...
+
+        # The Message Query Graph could be something like:
+        # "query_graph": {
+        #     "nodes": {
+        #         "type-2 diabetes": {"ids": ["MONDO:0005148"]},
+        #         "drug": {"categories": ["biolink:Drug"]}
+        #     },
+        #     "edges": {
+        #         "treats": {
+        #             "subject": "drug",
+        #             "predicates": ["biolink:treats"],
+        #             "object": "type-2 diabetes"
+        #         }
+        #     }
+        # }
+        #
+        # with a Response Message Knowledge Graph
+        # dictionary with 'nodes' and 'edges':
+        #
+        # "knowledge_graph": {
+        #     "nodes": ...,
+        #     "edges": ...
+        # }
+        knowledge_graph: Dict = message["knowledge_graph"]
+
+        # In the Knowledge Graph:
+        #
+        #     "nodes": {
+        #         "MONDO:0005148": {"name": "type-2 diabetes"},
+        #         "CHEBI:6801": {"name": "metformin", "categories": ["biolink:Drug"]}
+        #     }
+        #
+        # Check for case 'subject_id' and 'object_id',
+        # with expected categories, in nodes catalog
+        nodes: Dict = knowledge_graph["nodes"]
+        subject_id = case["subject_id"] if "subject_id" in case else case["subject"]
+        if not self.case_node_found("subject", subject_id, case, nodes):
+            # 'subject' node not found?
+            return False
+
+        object_id = case["object_id"] if "object_id" in case else case["object"]
+        if not self.case_node_found("object", object_id, case, nodes):
+            # 'object' node not found?
+            return False
+
+        # In the Knowledge Graph:
+        #
+        #     "edges": {
+        #         "df87ff82": {
+        #             "subject": "CHEBI:6801",
+        #             "predicate": "biolink:treats",
+        #             "object": "MONDO:0005148"
+        #         }
+        #     }
+        #
+        # Check in the edges catalog for an edge containing
+        # the case 'subject_id', 'predicate' and 'object_id'
+        edges: Dict = knowledge_graph["edges"]
+
+        predicate = case["predicate"]
+
+        validator = BiolinkValidator(TRAPIGraphType.Knowledge_Graph, biolink_version=self.biolink_version)
+        inverse_predicate = validator.get_inverse_predicate(predicate)
+
+        edge_id_found: Optional[str] = None
+        for edge_id, edge in edges.items():
+            # Note: this edge search could be arduous on a big knowledge graph?
+            if edge["subject"] == subject_id and \
+                    edge["predicate"] == predicate and \
+                    edge["object"] == object_id:
+                edge_id_found = edge_id
+                break
+            elif edge["subject"] == object_id and \
+                    edge["predicate"] == inverse_predicate and \
+                    edge["object"] == subject_id:
+                # observation of the inverse edge is also counted as a match?
+                edge_id_found = edge_id
+                break
+
+        if edge_id_found is None:
+            # Test case S--P->O edge not found?
+            return False
+
+        results: List = message["results"]
+        if not self.case_result_found(subject_id, object_id, edge_id_found, results, trapi_version):
+            # Some components of test case S--P->O edge
+            # NOT bound within any Results?
+            return False
+
+        # By this point, the case data assumed to be
+        # successfully validated in the TRAPI Response?
+        return True